deploy/new_panel/requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-socketio
aiofiles
requests
pydantic-settings
